LayoutTests/fast/scrolling/mac/smooth-scroll-with-overflow-hidden-and-layout-expected.txt

PASS scrollContainer.scrollLeft is 500
PASS successfullyParsed is true

TEST COMPLETE
